Move common fake API into separate files
